pyproject.toml
[tool.poetry]
name = "internet_speedtest_pi"
version = "0.1.0"
description = "Python package to measure the current internet speed and write it to a log file."
license = "MIT"
authors = ["Kilian Helfenbein <k.helfenbein@posteo.de>"]
readme = "README.rst"
homepage = "https://github.com/khelfen/internet-speedtest-pi"
repository = "https://github.com/khelfen/internet-speedtest-pi"
documentation = "https://khelfen.github.io/internet-speedtest-pi/"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
include = [
    "docs",
    "tests",
    "*.rst",
    "LICENSE",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/khelfen/internet-speedtest-pi/issues"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
dynaconf = "^3.1.7"
speedtest-cli = "^2.1.3"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
codecov = "^2.1.11"
coverage = "^5.5"
flake8 = "^3.9.1"
isort = "^5.8.0"
mypy = "^0.812"
pre-commit = "^2.12.1"
pylint = "^2.8.2"
pytest = "^6.2.4"
pytest-cov = "^2.11.1"
pytest-xdist = "^2.2.1"
Sphinx = "^3.5.4"
sphinx-autoapi = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
